EIT recording timing
This patch includes all the changes required for EIT recording timing on Freeview and Freesat.
It has been modified from the previous commits so that a recording only matches by CRID.
If there is not CRID (programme ID) present in the EIT data it will not match.

This uses the EIT present/following to mark the beginning and end of a recording.  A
bookmark is inserted when a programme becomes "present" from "following".  If a
programme is still "present" at the end of the scheduled time the recording will
continue.